added spring gem to spead up tests. Moved classes to services folder since these are technically service objects. Refactored tests
